Add async Select and Consume
