+++ Maybe.Test/MaybeTests.cs
@@ -1,5 +1,6 @@
-using FluentAssertions;
 using System;
+using System.Threading.Tasks;
+using FluentAssertions;
 using Xunit;
 
 namespace ZBRA.Maybe.Test
@@ -69,107 +70,6 @@ namespace ZBRA.Maybe.Test
         public void OrThrow_NullArgument_ShouldThrow()
         {
             Action subject = () => 1.ToMaybe().OrThrow(null);
-
-            subject.Should().ThrowExactly<ArgumentNullException>();
-        }
-
-        [Fact]
-        public void Consume_WithValue_ShouldExecuteAction()
-        {
-            string result = null;
-            var expected = 1;
-
-            expected.ToMaybe()
-                .Consume(i => result = i.ToString());
-
-            result.Should().Be(expected.ToString());
-        }
-
-        [Fact]
-        public void Consume_WithNoValue_ShouldNotExecuteAction()
-        {
-            var result = "a";
-
-            Maybe<string>.Nothing
-                .Consume(i => result = "b");
-
-            result.Should().Be("a");
-        }
-
-        [Fact]
-        public void Consume_NullArgument_ShouldThrow()
-        {
-            Action subject = () => 1.ToMaybe().Consume(null);
-
-            subject.Should().ThrowExactly<ArgumentNullException>();
-        }
-
-        [Theory]
-        [InlineData(null, null, null)]
-        [InlineData(null, 2, null)]
-        [InlineData(2, null, null)]
-        [InlineData(1, 2, "3")]
-        public void Zip_WithTransformer_ShouldZipValues(int? value, double? otherValue, string expected)
-        {
-            static string transformer(int v, double o) => (v + o).ToString();
-
-            var result = value.ToMaybe()
-                .Zip(otherValue.ToMaybe(), transformer);
-
-            result.Should().Be(expected.ToMaybe());
-        }
-
-        [Theory]
-        [InlineData(null, null, null)]
-        [InlineData(null, 2, null)]
-        [InlineData(2, null, null)]
-        [InlineData(1, 2, "3")]
-        public void Zip_WithMaybeTransformer_ShouldZipValues(int? value, double? otherValue, string expected)
-        {
-            static Maybe<string> transformer(int v, double o) => (v + o).ToString().ToMaybe();
-
-            var result = value.ToMaybe()
-                .Zip(otherValue.ToMaybe(), transformer);
-
-            result.Should().Be(expected.ToMaybe());
-        }
-
-        [Fact]
-        public void Zip_NullArgument_ShouldThrow()
-        {
-            Action subject = () => 1.ToMaybe().Zip(Maybe<int>.Nothing, (Func<int, int, int>)null);
-
-            subject.Should().ThrowExactly<ArgumentNullException>();
-        }
-
-        [Fact]
-        public void Zip_MaybeNullArgument_ShouldThrow()
-        {
-            Action subject = () => 1.ToMaybe().Zip(Maybe<int>.Nothing, (Func<int, int, Maybe<int>>)null);
-
-            subject.Should().ThrowExactly<ArgumentNullException>();
-        }
-
-        [Theory]
-        [InlineData(null, null, null)]
-        [InlineData(null, 2, null)]
-        [InlineData(2, null, null)]
-        [InlineData(1, 2, "3")]
-        public void ZipConsume_WhenBothValuesExist_ShouldExecuteAction(int? value, double? otherValue, string expected)
-        {
-            string result = null;
-            void action(int v, double o) => result = (v + o).ToString();
-
-            value.ToMaybe()
-                .ZipAndConsume(otherValue.ToMaybe(), action);
-
-            result.Should().Be(expected);
-        }
-
-        [Fact]
-        public void ZipAndConsume_NullArgument_ShouldThrow()
-        {
-            Action subject = () => 1.ToMaybe().ZipAndConsume(Maybe<int>.Nothing, null);
 
             subject.Should().ThrowExactly<ArgumentNullException>();
         }
